api/requirements.txt
uvicorn==0.34.0
python-dotenv==1.0.1
fastapi==0.115.6
firebase-admin==6.6.0
spotipy==2.24.0
pandas==2.2.3
mistralai==1.2.5
requests==2.32.3